# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zak Ecosystem AI - CRM Omnicanale API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /automazioni/log:
    get:
      operationId: listAutomazioniLog
      tags: [automazioni]
      summary: Get automation execution log
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomazioneLog"
...
  /automazioni/config/{chiave}:
    patch:
      operationId: updateAutomazioneConfig
      tags: [automazioni]
      summary: Update an automation config value such as timing, toggles, or event-type segments
      parameters:
        - name: chiave
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomazioneConfigUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomazioneConfig"
  /automazioni/trigger:
    post:
      operationId: triggerAutomazione
      tags: [automazioni]
      summary: Manually trigger an automation job using the current segmentation config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerManualInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerManualResponse"
...
components:
  schemas:
...
    AutomazioneLog:
      type: object
      required: [id, tipo, stato, data_esecuzione]
      properties:
        id:
          type: string
        tipo:
          type: string
        contatto_id:
          type: ["string", "null"]
        contatto_nome:
          type: ["string", "null"]
        stato:
          type: string
        messaggio:
          type: ["string", "null"]
        data_esecuzione:
          type: string
    AutomazioneConfig:
      type: object
      required: [chiave, valore, descrizione, aggiornato_il]
      properties:
        chiave:
          type: string
        valore:
          type: string
        descrizione:
          type: ["string", "null"]
        aggiornato_il:
          type: string
    AutomazioneConfigUpdate:
      type: object
      required: [valore]
      properties:
        valore:
          type: string
...
    TriggerManualInput:
      type: object
      required: [tipo]
      properties:
        tipo:
          type: string
    TriggerManualResponse:
      type: object
      required: [eseguiti]
      properties:
        eseguiti:
          type: integer
        dettagli:
          type: array
          items:
            type: string
...